@@ -10,8 +10,8 @@ package cn.com.heaton.blelibrary.ble;
 import android.bluetooth.le.ScanFilter;
 import android.content.Context;
 import android.os.Build;
-import android.support.annotation.IntRange;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.IntRange;
+import androidx.annotation.RequiresApi;
 
 import java.util.UUID;
 
