@@ -1,7 +1,7 @@
 package cn.com.heaton.blelibrary.ble.callback;
 
 import android.os.Build;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.RequiresApi;
 
 import cn.com.heaton.blelibrary.ble.model.ScanRecord;
 
